// src/api/models/accounts-password-reset-request-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AccountsPasswordResetRequestCreateRequest
 */
export interface AccountsPasswordResetRequestCreateRequest {
    /**
     * Email address associated with the user\'s account
     * @type {string}
     * @memberof AccountsPasswordResetRequestCreateRequest
     */
    'email': string;
}