// src/api/models/accounts-password-reset-verify-otp-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AccountsPasswordResetVerifyOtpCreateRequest
 */
export interface AccountsPasswordResetVerifyOtpCreateRequest {
    /**
     * User\'s email address
     * @type {string}
     * @memberof AccountsPasswordResetVerifyOtpCreateRequest
     */
    'email': string;
    /**
     * 6-digit OTP sent via email
     * @type {string}
     * @memberof AccountsPasswordResetVerifyOtpCreateRequest
     */
    'otp': string;
}