// src/api/models/accounts-password-reset-reset-password-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AccountsPasswordResetResetPasswordCreateRequest
 */
export interface AccountsPasswordResetResetPasswordCreateRequest {
    /**
     * User\'s email address
     * @type {string}
     * @memberof AccountsPasswordResetResetPasswordCreateRequest
     */
    'email': string;
    /**
     * 6-digit OTP sent to email
     * @type {string}
     * @memberof AccountsPasswordResetResetPasswordCreateRequest
     */
    'otp': string;
    /**
     * New password to be set
     * @type {string}
     * @memberof AccountsPasswordResetResetPasswordCreateRequest
     */
    'new_password': string;
}